Commit 3

+++ studentdata/src/main/java/edu/mum/cs425/studentdata/repository/ClassRoomRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ClassRoomRepository extends JpaRepository<Classroom, Long> {
+public interface ClassRoomRepository extends JpaRepository<Classroom, Integer> {
 }
